chore: refactor ContextHolder implementation
move Context and ContextHolder to the root library package.
initialize ContextHolder to InheritableThreadLocal if configuration property is provided.
Add support for additional configuration property into LoggingConfig.
