+++ google-cloud-logging/src/test/java/com/google/cloud/logging/ContextTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.logging.context;
+package com.google.cloud.logging;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -22,7 +22,6 @@ import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-import com.google.cloud.logging.HttpRequest;
 import org.junit.Test;
 import org.threeten.bp.Duration;
